Remove code which added a default reason when the first one wasn't expected
This code was probably leftover from a previous implementation which wasn't
collecting reasons properly. Now, it's possible that the first dependency
we see is actually a constraint, so the test didn't make sense.

It also fixes a rendering issue where multiple blank lines were added in
case of rendering extra sections.